Major interface change.
Now methods that used to return True on success and False on failure now have no
return value and raise an exception on failure.

This has the side effect of making visible certain failures that were previously ignored
(because TotalConnectClient itself sometimes did not check the return value of its own
methods).